changed packets.h and made progress on ignition fw

+++ firmware/comms-router/packets.h
@@ -84,7 +84,7 @@ typedef struct __attribute__((packed)) {
     uint8_t     bank;
     uint8_t     state;
 
-    uint16_t    mcu_temp;
+    float    mcu_temp;
 
     float    psu_voltage;
     float    firing_bus_voltage;
